# services/api-gateway/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/accounts/emails/{email}": {
      "get": {
        "operationId": "getAccountByEmail",
        "summary": "Get an user account by account email",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDto"
                }
              }
            }
          }
        },
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/accounts/{id}": {
      "get": {
        "operationId": "getAccountById",
        "summary": "Get an user account by account id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDto"
                }
              }
            }
          }
        },
        "tags": [
          "Accounts"
        ]
      }
    },
...
    "/api/accounts/admin/emails/{email}": {
      "get": {
        "operationId": "adminGetAccountByEmail",
        "summary": "Get an user account by account email with admin authority",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDto"
                }
              }
            }
          }
        },
        "tags": [
          "Accounts"
        ]
      }
    },
...
    "/api/fields/create": {
      "post": {
        "operationId": "createField",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFieldBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Fields"
        ]
      }
    },
    "/api/fields/{id}": {
      "get": {
        "operationId": "getFieldById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Fields"
        ]
      }
    },
    "/api/fields/update/{id}": {
      "put": {
        "operationId": "updateFieldById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFieldBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Fields"
        ]
      }
    },
    "/api/fields/delete/{id}": {
      "delete": {
        "operationId": "deleteFieldById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Fields"
        ]
      }
    },
...
    "/api/profiles/create": {
      "post": {
        "operationId": "registerProfile",
        "summary": "Register a new account's profile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfileBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            }
          }
        },
        "tags": [
          "Profiles"
        ]
      }
    },
    "/api/profiles/{id}": {
      "get": {
        "operationId": "getProfileById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Profiles"
        ]
      }
    },
    "/api/profiles/accounts/{accountId}": {
      "get": {
        "operationId": "getProfileByAccountId",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Profiles"
        ]
      }
    },
    "/api/profiles/update/accounts/{accountId}": {
      "put": {
        "operationId": "updateProfileByAccountId",
        "summary": "Update an account's profile by account id",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            }
          }
        },
        "tags": [
          "Profiles"
        ]
      }
    },
    "/api/projects/create": {
      "post": {
        "operationId": "createProject",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/projects/{id}": {
      "get": {
        "operationId": "getProjectById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/projects/update/{id}": {
      "put": {
        "operationId": "updateProjectById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
...
    "/api/specializations/create": {
      "post": {
        "operationId": "createSpecialization",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpecializationBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Specializations"
        ]
      }
    },
    "/api/specializations/{id}": {
      "get": {
        "operationId": "getSpecializationById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Specializations"
        ]
      }
    },
    "/api/specializations/update/{id}": {
      "put": {
        "operationId": "updateSpecializationById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSpecializationBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Specializations"
        ]
      }
    },
    "/api/specializations/delete/{id}": {
      "delete": {
        "operationId": "deleteSpecializationById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Specializations"
        ]
      }
    },
...
  },
  "info": {
    "title": "SV Connect API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "AccountDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "emailVerified": {
            "type": "boolean"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "email",
          "emailVerified",
          "role"
        ]
      },
...
      "ConnectSpecializationBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "CreateFieldBody": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "specializations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectSpecializationBody"
            }
          }
        },
        "required": [
          "title"
        ]
      },
      "DisconnectSpecializationBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "UpdateManySpecializationsRelationBody": {
        "type": "object",
        "properties": {
          "connect": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectSpecializationBody"
            }
          },
          "disconnect": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DisconnectSpecializationBody"
            }
          }
        }
      },
      "UpdateFieldBody": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "specializations": {
            "$ref": "#/components/schemas/UpdateManySpecializationsRelationBody"
          }
        }
      },
      "ConnectStudentBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "ConnectAccountBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "CreateProfileBody": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "headline": {
            "type": "object"
          },
          "summary": {
            "type": "object"
          },
          "pictureUrl": {
            "type": "object"
          },
          "backgroundUrl": {
            "type": "object"
          },
          "account": {
            "$ref": "#/components/schemas/ConnectAccountBody"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "account"
        ]
      },
      "ProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "headline": {
            "type": "object"
          },
          "summary": {
            "type": "object"
          },
          "pictureUrl": {
            "type": "object"
          },
          "backgroundUrl": {
            "type": "object"
          },
          "accountId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "firstName",
          "lastName",
          "headline",
          "summary",
          "pictureUrl",
          "backgroundUrl",
          "accountId"
        ]
      },
      "UpdateProfileBody": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "headline": {
            "type": "object"
          },
          "summary": {
            "type": "object"
          },
          "pictureUrl": {
            "type": "object"
          },
          "backgroundUrl": {
            "type": "object"
          }
        }
      },
      "ConnectFieldBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "CreateProjectBody": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "field": {
            "$ref": "#/components/schemas/ConnectFieldBody"
          },
          "specializations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectSpecializationBody"
            }
          },
          "student": {
            "$ref": "#/components/schemas/ConnectStudentBody"
          }
        },
        "required": [
          "title",
          "summary",
          "field",
          "specializations",
          "student"
        ]
      },
      "UpdateOneFieldRelationBody": {
        "type": "object",
        "properties": {
          "connect": {
            "$ref": "#/components/schemas/ConnectFieldBody"
          },
          "disconnect": {
            "type": "boolean"
          }
        }
      },
      "UpdateProjectBody": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "field": {
            "$ref": "#/components/schemas/UpdateOneFieldRelationBody"
          },
          "specializations": {
            "$ref": "#/components/schemas/UpdateManySpecializationsRelationBody"
          }
        }
      },
      "CreateSpecializationBody": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectFieldBody"
            }
          }
        },
        "required": [
          "title"
        ]
      },
      "DisconnectFieldBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "UpdateManyFieldsRelationBody": {
        "type": "object",
        "properties": {
          "connect": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectFieldBody"
            }
          },
          "disconnect": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DisconnectFieldBody"
            }
          }
        }
      },
      "UpdateSpecializationBody": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "fields": {
            "$ref": "#/components/schemas/UpdateManyFieldsRelationBody"
          }
        }
      },
...
    }
  }
}